Move source files to new library "BlueMystical.Forms.DarkModeNet"
