+++ Dark-Mode-Forms/Properties/Resources.resx
