+++ Dark-Mode-Forms/Resources/forward_16x16.png
