+++ DarkModeForms.sln
@@ -1,8 +1,11 @@
-﻿Microsoft Visual Studio Solution File, Format Version 12.00
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34622.214
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DarkModeForms", "DarkModeForms.csproj", "{3972405E-252E-4EAE-97B7-204957260480}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DarkModeForms", "Dark-Mode-Forms\DarkModeForms.csproj", "{3972405E-252E-4EAE-97B7-204957260480}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMystical.Forms.DarkModeNet", "BlueMystical.Forms.DarkModeNet\BlueMystical.Forms.DarkModeNet.csproj", "{0C318329-3424-4F53-888E-89F1FBE0724A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -14,6 +17,10 @@ Global
 		{3972405E-252E-4EAE-97B7-204957260480}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3972405E-252E-4EAE-97B7-204957260480}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3972405E-252E-4EAE-97B7-204957260480}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C318329-3424-4F53-888E-89F1FBE0724A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0C318329-3424-4F53-888E-89F1FBE0724A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C318329-3424-4F53-888E-89F1FBE0724A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0C318329-3424-4F53-888E-89F1FBE0724A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
